% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{Calculate Prevalence}
\usage{
calculate_prevalence(data, group_var, count_var)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{group_var}{The column of thecategorical variable used for grouping.}

\item{count_var}{The column of the categorical variable whose prevalence is to be calculated related to group_var.}
}
\value{
A data frame containing the prevalence calculation, including the count var, the count (the number of group_var count_var is represented in) and proportion.
}
\description{
Calculate the prevalence of a categorical variable related to another categorical variable in a dataset.
}
\examples{
library(magrittr)
data <- data.frame(group_var = c("A", "B", "A", "B"),
                   count_var = c("F1", "G1", "G1", "G1"))
prevalence <- data \%>\% calculate_prevalence(group_var, count_var )

}
