% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_prevalence.R
\name{filter_by_prevalence}
\alias{filter_by_prevalence}
\title{Filter data by Prevalence}
\usage{
filter_by_prevalence(data, group_var, count_var, limit, at_least = TRUE)
}
\arguments{
\item{data}{A data frame containing the data to be filtered.}

\item{group_var}{The name of the categorical variable used for grouping.}

\item{count_var}{The name of the categorical variable whose prevalence is to be calculated related to group_var.}

\item{limit}{The limit value for the prevalence. Must be a valid number lower than 1.}

\item{at_least}{Logical value indicating whether to filter for values of prevalence at least greater than or equal to the limit (TRUE) or values less than the limit (FALSE). Default is TRUE.}
}
\value{
A data frame containing the filtered data based on prevalence.
}
\description{
Filter data based on the prevalence of a categorical variable related to another categorical variable, considering a specified limit and direction.
}
\examples{
library(dplyr)
data <- data.frame(group_var = c("A", "B", "A", "B", "C"),
                    count_var = c("G1", "G2", "G2", "G2", "G2"))
 result <- data \%>\% filter_by_prevalence(group_var, count_var, limit = 0.5, at_least = FALSE)

}
