% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_prevalence.R
\name{filter_by_prevalence}
\alias{filter_by_prevalence}
\title{Filter Dataset by Prevalence}
\usage{
filter_by_prevalence(dataset, group_var, count_var, limit, at_least = TRUE)
}
\arguments{
\item{dataset}{A data frame containing the dataset to be filtered.}

\item{group_var}{The name of the categorical variable used for grouping.}

\item{count_var}{The name of the categorical variable used for counting.}

\item{limit}{The limit value for the prevalence. Must be a valid number lower than 1.}

\item{at_least}{Logical value indicating whether to filter for values at least greater than or equal to the limit (TRUE) or values less than the limit (FALSE). Default is TRUE.}
}
\value{
A data frame containing the filtered dataset based on the prevalence filtering.
}
\description{
Filter a dataset based on the prevalence of a categorical variable related to another categorical variable, considering a specified limit and direction.
}
\examples{
library(dplyr)
data <- data.frame(group_var = c("A", "B", "A", "B"),
                   count_var = c("F1", "F1", "G1", "G1"))
filtered_data <- data \%>\% pRevalence::calculate_prevalence(group_var = "group_var", count_var = "count_var")
filter_by_prevalence(data, group_var = "group_var", count_var = "count_var", limit = 0.4, at_least = TRUE)

}
